+++ api/main.py
@@ -1,23 +1,18 @@
 # 📦 Import necessary modules and packages
-from enum import Enum
-import re
-import time
-import requests
+from manufacturers.brighter_bins import get_bb_readings, update_bb_cache
+from manufacturers.tekelek import  get_tkl_readings, update_tkl_cache
 from services.google_routes import RouteRequest, get_optimized_routes_response
 from services.mail import AlertEmailSchema, EmailSchema, get_email_msg, get_fm
 from utils import filter_nulls
 
-import gspread
 from dotenv import load_dotenv
 
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi_utils.tasks import repeat_every
 
-from pydantic import BaseModel
 import os
-from datetime import datetime
-from typing import Optional, List, Dict, Union
+
 
 from starlette.responses import JSONResponse
 
@@ -26,17 +21,6 @@ load_dotenv()
 
 # 🛠️ Get environment configurations
 env = os.environ.get("ENVIRONMENT")
-bb_email = os.environ.get("BB_EMAIL")
-bb_password = os.environ.get("BB_PASSWORD")
-tkl_client = os.environ.get("TEKELEK_CLIENT")
-tkl_password = os.environ.get("TEKELEK_PASSWORD")
-tkl_secret = os.environ.get("TEKELEK_TOKEN")
-tkl_username = os.environ.get("TEKELEK_USERNAME")
-
-# 📚 Authenticate with Google Sheets service account
-sa = gspread.service_account(filename="google_sheets_sa_key.json")
-
-goog_sheet_date_format = "%m/%d/%Y %H:%M:%S"
 
 # 🚀 Initialize FastAPI application
 app = FastAPI()
@@ -59,509 +43,32 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-
-# 📊 Enum to represent sensor types
-class SensorType(Enum):
-    LIQUID_BIN_LEVEL = "liquid bin level"
-    SOLID_BIN_LEVEL = "solid bin level"
-
-
-# 📝 Model to represent a basic sensor
-class BasicSensor(BaseModel):
-    id: str
-    sensor_type: SensorType
-    fill_level: int | None
-    lat: float | None
-    long: float | None
-    manufacturer: str
-    bin_name: str | None
-    address_line1: str | None
-    address_line2: str | None
-    city: str | None
-    province: str | None
-    postal_code: str | None
-    group: str | None
-    bin_type: str | None
-    material_type: str
-    asset_tag: str
-    bin_volume: str
-    fill_level_last_collected: int | None
-    fill_level_alert: int | None
-    temperature_alert: int | None
-    illegal_dumping_alert: bool | None
-    contamination_alert: bool | None
-    last_collected: datetime | None
-
-
-
-# 📝 Model to represent a BrighterBinsSensorReading
-class BrighterBinsSensorReading(BaseModel):
-    epoch_ms: int | None
-    rawDistance: int | None
-    temperature: int | None
-    batteryLevel: int | None
-    pickUpEvent: bool | None
-    fillLevel: int
-    fillError: bool | None
-    snr: int | None
-    rssi: int | None
-    lat: float | None
-    long: float | None
-
-
-# 📝 Model to represent a BrighterBinsSensor
-class BrighterBinsSensor(BaseModel):
-    id: str
-    row_id: int | None
-    readings: list[BrighterBinsSensorReading] | None
-    is_extended_uplink: int | None
-    manufacturer: str | None
-    # fields required by BasicSensor
-    id: str
-    fill_level: int | None
-    lat: float | None
-    long: float | None
-    bin_name: str | None
-    address: str | None
-    city: str | None
-    province: str | None
-    postal_code: str | None
-    group: str | None
-    bin_type: str
-    material_type: str
-    asset_tag: str
-    bin_volume: str
-    fill_level_last_collected: int | None
-    fill_level_alert: int | None
-    temperature_alert: int | None
-    illegal_dumping_alert: bool | None
-    contamination_alert: bool | None
-    last_collected: datetime | None
-
-# 📝 Model to represent a TeklekSensor
-class TekelekSensor(BaseModel):
-    ModemSerialNo: Optional[str]
-    Shape: Optional[str]
-    TheoreticalCapacity: Optional[float]
-    SafePercentage: Optional[int]
-    NominalCapacity: Optional[float]
-    Diameter: Optional[float]
-    Bund: Optional[bool]
-    Height: Optional[float]
-    Width: Optional[float]
-    Length: Optional[float]
-    SensorOffset: Optional[float]
-    OutletHeight: Optional[float]
-    Make: Optional[str]
-    Model: Optional[str]
-    Substance: Optional[str]
-    Material: Optional[str]
-    AccountNo: Optional[str]
-    DateCreated: Optional[str]
-    DateLastModified: Optional[str]
-    Note: Optional[str]
-    Location: Optional[dict] = None
-    ContainerType: Optional[str]
-    OrderTriggerPercentage: Optional[str]
-    Group: Optional[str] = (None,)
-    # fields required by BasicSensor
-    id: str
-    fill_level: int | None
-    lat: float | None
-    long: float | None
-    bin_name: str | None
-    address_line1: str | None
-    address_line2: str | None
-    city: str | None
-    province: str | None
-    postal_code: str | None
-    group: str | None
-    bin_type: str
-    material_type: str
-    asset_tag: str
-    bin_volume: str
-    fill_level_last_collected: int | None
-    fill_level_alert: int | None
-    temperature_alert: int | None
-    illegal_dumping_alert: bool | None
-    contamination_alert: bool | None
-    last_collected: datetime | None
-
-
-# 🔧 Function to convert TekelekSensor to BasicSensor
-def tkl_to_bs(sensor: TekelekSensor) -> BasicSensor:
-    #lat, long = None, None  # Default values in case of errors
-
-    # try:
-    #     location = sensor.Location
-
-    #     if location:
-    #         well_known_text = location.get("Geography", {}).get("WellKnownText", None)
-
-    #         if well_known_text and well_known_text.startswith("POINT"):
-    #             pattern = r"POINT \((-?\d+\.\d+) (-?\d+\.\d+)\)"
-    #             match = re.search(pattern, well_known_text)
-
-    #             if match:
-    #                 lat = float(match.group(1))
-    #                 long = float(match.group(2))
-
-    # except (AttributeError, TypeError, ValueError) as e:
-    #     print(f"Error while processing location data: {e}")
-
-    bs = BasicSensor(
-        id=sensor.id,
-        sensor_type=SensorType.LIQUID_BIN_LEVEL,
-        fill_level=sensor.fill_level,
-        lat=sensor.lat,
-        long=sensor.long,
-        manufacturer='Tekele',
-        bin_name=sensor.bin_name,
-        address_line1=sensor.address_line1,
-        address_line2=sensor.address_line2,
-        city=sensor.city,
-        province=sensor.province,
-        postal_code=sensor.postal_code,  
-        group=sensor.group,
-        bin_type=sensor.bin_type,
-        material_type=sensor.material_type,
-        asset_tag=sensor.asset_tag,
-        bin_volume=sensor.bin_volume,
-        fill_level_last_collected=sensor.fill_level_last_collected, 
-        fill_level_alert=sensor.fill_level_alert,  
-        temperature_alert=sensor.temperature_alert,  
-        illegal_dumping_alert=sensor.illegal_dumping_alert,  
-        contamination_alert=sensor.contamination_alert,  
-        last_collected=sensor.last_collected  
-    )
-    return bs
-
-
-
-# 🔧 Function to convert TeklekSensor dictionary to BasicSensor dictionary
-# def tkl_dict_to_bs_dict(sensors: List[Dict[str, Union[str, int, float, None]]]) -> dict:
-#     bs_dict = {}
-#     # loop to iterate through each obj in the TekelekSensor list
-#     for sensor_data in sensors:
-#         try:
-#             # create a TekelekSensor object (sensor_obj) by passing the dictionary sensor_data as keyword argument
-#             sensor_obj = TekelekSensor(**sensor_data)
-
-#             # print(f"sensor_obj : {sensor_obj }")
-#             # convert the TekelekSensor into a BasicSensor and store it in the bs_dict dictionary.
-#             bs_dict[sensor_obj.ModemSerialNo] = tkl_to_bs(sensor_obj)
-#         except (ValueError, KeyError, TypeError) as e:
-#             print(f"Error processing sensor {sensor_obj.ModemSerialNo}: {e}")
-#             # Optionally, continue to the next iteration
-#             continue
-#     return bs_dict
-
-
-# 🔧 Function to convert BrighterBinsSensor to BasicSensor with latest reading
-def bb_to_bs(
-    sensor: BrighterBinsSensor,
-) -> BasicSensor:
-    return BasicSensor(
-        id=sensor.id,
-        sensor_type=SensorType.SOLID_BIN_LEVEL,
-        fill_level=sensor.readings[-1].fillLevel # get the latest reading
-        if (sensor.readings and len(sensor.readings) > 0)
-        else None,
-        lat=sensor.lat,
-        long=sensor.long,
-        manufacturer="BrighterBins",
-        bin_name=sensor.bin_name,
-        address_line1=sensor.address,
-        address_line2=None,
-        city=sensor.city,
-        province=sensor.province,
-        postal_code=sensor.postal_code,
-        group=sensor.group,
-        bin_type=sensor.bin_type,
-        material_type=sensor.material_type,
-        bin_volume=sensor.bin_volume,
-        asset_tag=sensor.asset_tag,
-        fill_level_last_collected=sensor.fill_level_last_collected,
-        fill_level_alert=sensor.fill_level_alert,
-        temperature_alert=sensor.temperature_alert,
-        illegal_dumping_alert=sensor.illegal_dumping_alert,
-        contamination_alert=sensor.contamination_alert,
-        last_collected=sensor.last_collected,
-    )
-
-
-# 🔧 Function to fetch BrighterBins API token
-def get_brighterbins_token():
-    global bb_email
-    global bb_password
-    url = "https://api.brighterbins.com/auth/login"
-    payload = {"email": bb_email, "password": bb_password}
-
-    try:
-        response = requests.request("POST", url, data=payload)
-        response_json = response.json()
-
-        # Check if the response contains the "token" key
-        if "token" in response_json:
-            return response_json["token"]
-        elif "secode" in response_json:
-            print(
-                "Error in get_brighterbins_token: An error occurred while trying to fetch the token."
-            )
-            print("Error:", response_json["message"])
-            return None
-        else:
-            # Handle unexpected response format
-            print(
-                "Error in get_brighterbins_token: An unexpected response format was received."
-            )
-            print("Response:", response_json)
-            return None
-
-    except requests.exceptions.RequestException as e:
-        # Handle connection or request errors
-        print("Error in get_brighterbins_token: A request error occurred.")
-        print("Error:", e)
-        return None
-
-    except ValueError as e:
-        # Handle JSON decoding error
-        print("Error in get_brighterbins_token: A JSON decoding error occurred.")
-        print("Error:", e)
-        return None
-
-
-# 🔧 Function to fetch Tekelek API token
-def get_tekelek_token():
-    url = "https://phoenixidentityserverprod.azurewebsites.net/connect/token"
-    payload = {
-        "client_id": tkl_client,
-        "client_secret": tkl_secret,
-        "grant_type": "password",
-        "scope": "openid",
-        "username": tkl_username,
-        "password": tkl_password,
-    }
-
-    try:
-        # Use the generic request function to make the request
-        response_json = make_http_request(url, method="POST", data=payload)
-
-        # Check if the response contains the "access_token" key
-        if "access_token" in response_json:
-            return response_json["access_token"]
-        elif "error" in response_json:
-            print(
-                "Error in get_tekelek_token: An error occurred while trying to fetch the token."
-            )
-            print("Error:", response_json["error"])
-            return None
-        else:
-            # Handle unexpected response format
-            print(
-                "Error in get_tekelek_token: An unexpected response format was received."
-            )
-            print("Response:", response_json)
-            return None
-
-    except Exception as e:
-        # Handle any unexpected errors
-        print("Error in get_tekelek_token:", e)
-        return None
-
-
-# 🔧 Function for making HTTP requests
-def make_http_request(url, method="GET", headers=None, params=None, data=None):
-    try:
-        response = requests.request(
-            method, url, headers=headers, params=params, data=data
-        )
-        response.raise_for_status()  # Check for HTTP errors
-        return response.json()
-    except requests.exceptions.RequestException as request_err:
-        print("An error occurred while making the request:", request_err)
-        return None
-    except ValueError as value_err:
-        print("An error occurred while parsing JSON response:", value_err)
-        return None
-    except Exception as e:
-        print("An unexpected HTTP request error occurred:", e)
-        return None
-
-
 # Caches for sensor data
 bb_cache = {}
 tkl_cache = {}
 last_run_timestamp = 0
-
 
 #  🤖 Event handler: Populate Tekelek data in tkl_cache with initial data on start up
 @app.on_event("startup")
 def set_tkl_cache():
     global tkl_cache
     try:
-        print("Tekelek sensor data cache: fetching sensor data...")
-        # get api token
-        tekelek_api_token = get_tekelek_token() 
-
-        # Define the base URL
-        base_url = "https://phoenixapiprod.azurewebsites.net/api/"
-
-        # Make a request to get all sensor data from tank records
-        # tanks_url = base_url + "tanks"
-        # tanks_response = make_http_request(
-        #     tanks_url,
-        #     method="GET",
-        #     headers={"Authorization": "Bearer " + tekelek_api_token},
-        # )
-        sheet = sa.open("Mississauga_mock_data")
-        worksheet = sheet.worksheet("Asset-Bin Information-TekelekTA")
-        records = worksheet.get_all_records()
-
-        if records:
-            # Iterate asset records to get additional information from related API endpoints
-            for index, record in enumerate(records):
-                id = record["Sensor ID"]
-                print("fetching Tekelek id", id) 
-                # get the latest sensor fill reading
-                latest_reading_url = base_url + "latestReading/" + str(id)
-                reading_response = make_http_request(
-                    latest_reading_url,
-                    method="GET",
-                    headers={"Authorization": "Bearer " + tekelek_api_token},
-                )
-
-                if not (reading_response):
-                    print("failed to get response from id ", id)
-
-                sensor = {
-                    "id": id,
-                    "row_id": index + 2,
-                    "bin_name": record["Asset - Name"],
-                    "address_line1": record["Address"],
-                    "address_line2": None,
-                    "city": record["City"],
-                    "province": record["Province"],
-                    "postal_code": record["Postal code"],
-                    "lat": record["Latitude"],
-                    "long": record["Longitude"],
-                    "bin_volume": record["Bin Volume"],
-                    "bin_type": record["Bin Type"],
-                    "material_type": record["Material/Waste Type"],
-                    "asset_tag": record["Addiontal Asset Tags"],
-                    "group": record["Group"],
-                    "fill_level_last_collected": record["Fill_level_last_collected"],
-                    "fill_level": reading_response["PercentFull"] if reading_response else None,
-                    "fill_level_alert": record["Fill_level_alert"],
-                    "temperature_alert": record["Temperature_alert"],
-                    "illegal_dumping_alert": record["Illegal_dumping_alert"].upper() == "YES",
-                    "contamination_alert": record["Contamination_alert"].upper() == "YES",
-                    "last_collected": datetime.strptime(record["Last_collected"], goog_sheet_date_format)
-                }
-                tkl_cache[id] = sensor
-            print("Tekelek sensor data cache: fetch complete")
+        tkl_cache = update_tkl_cache(tkl_cache)
 
     except Exception as e:
         print("An unexpected error occurred:", e)
 
 
-#  🤖 Event handler: Fetch BrighterBins historical data and update bb_cache hourly
-# The purpose is to maintain a historical record of sensor readings over time.
-# This allows the program to access and analyze past readings without repeatedly querying the API for the same data
-# event handler continues to run periodically, in the background, due to the @repeat_every decorator
 @app.on_event("startup")
 @repeat_every(seconds=60 * 60)  # Every 1 hour
-def update_bb_cache() -> None:
+def set_bb_cache() -> None:
     global bb_cache
     global last_run_timestamp
-    brighterbins_api_token = get_brighterbins_token()
-    url = "https://api.brighterbins.com/bins/timeseries"
 
-    # fully populate bb_cache if this is the first time this is being run
-    if last_run_timestamp == 0:
-        print("BrighterBins cache: Fetching initial data...")
-        readingsEndTime = round(time.time() * 1000)
-        # start around July 28, for no particular reason
-        readingsStartTime = 1690592310000
-        last_run_timestamp = readingsEndTime + 1
-        # get all the sensors data from the google sheet
-        sheet = sa.open("Mississauga_mock_data")
-        worksheet = sheet.worksheet("Asset-Bin Information-BBTA")
-        records = worksheet.get_all_records()
-
-        for index, record in enumerate(records):
-            id = record["Sensor ID"]
-            print("fetching BrighterBin id", id)
-            try:
-                response = requests.request(
-                    "POST",
-                    url,
-                    headers={"Authorization": "Bearer " + brighterbins_api_token},
-                    data={
-                        "from": readingsStartTime,
-                        "to": readingsEndTime,
-                        "id": id,
-                    },
-                ).json()
-                # Check if the response is successful
-                if response.get('success'):
-                    readings = (
-                        []
-                        if len(response["data"]["series"]) == 0
-                        else response["data"]["series"]
-                    )
-                else:
-                    print(f"Error fetching data for sensor {id}. Response: {response}")
-                    readings = ([{"fillLevel": -1}])
-                
-                sensor = {
-                    "id": id,
-                    "row_id": index + 2,
-                    "bin_name": record["Asset - Name"],
-                    "address": record["Address"],
-                    "city": record["City"],
-                    "province": record["Province"],
-                    "postal_code": record["Postal code"],
-                    "lat": record["Latitude"],
-                    "long": record["Longitude"],
-                    "bin_volume": record["Bin Volume"],
-                    "bin_type": record["Bin Type"],
-                    "material_type": record["Material/Waste Type"],
-                    "asset_tag": record["Addiontal Asset Tags"],
-                    "group": record["Group"],
-                    "readings": readings,
-                    "fill_level_last_collected": record["Fill_level_last_collected"],
-                    "fill_level_alert": record["Fill_level_alert"],
-                    "temperature_alert": record["Temperature_alert"],
-                    "illegal_dumping_alert": record["Illegal_dumping_alert"].upper() == "YES",
-                    "contamination_alert": record["Contamination_alert"].upper() == "YES",
-                    "last_collected": datetime.strptime(record["Last_collected"], goog_sheet_date_format)
-                }
-                bb_cache[id] = sensor
-            except Exception as e:
-                print(f"Error fetching data for sensor {id}. Exception: {e}")
-
-        print("Initial fetch complete")
-    else:
-        print("Fetching latest data...")
-        for id in bb_cache:
-            readingsEndTime = round(time.time() * 1000)
-            readingsStartTime = last_run_timestamp + 1
-            last_run_timestamp = readingsEndTime + 1
-            response = requests.request(
-                "POST",
-                url,
-                headers={"Authorization": "Bearer " + brighterbins_api_token},
-                data={
-                    "from": readingsStartTime,
-                    "to": readingsEndTime,
-                    "id": id,
-                },
-            ).json()
-            bb_cache[id]["readings"] = (
-                bb_cache[id]["readings"] + response["data"]["series"]
-            )
-        print("Latest fetch complete")
+    try:
+        bb_cache, last_run_timestamp = update_bb_cache(bb_cache, last_run_timestamp)
+    except Exception as e:
+        print("An unexpected error occurred:", e)
 
 
 # 🚀 API endpoint: Get the latest readings from all types of sensors
@@ -570,20 +77,8 @@ def get_latest_readings():
     global bb_cache
     global tkl_cache
 
-    bb_readings = []
-    for bb_id, bb_sensor in bb_cache.items():
-        bb_obj = BrighterBinsSensor(**bb_sensor)
-        print(565, bb_sensor['lat'], bb_obj.lat)
-        bb_readings.append(
-            bb_to_bs(bb_obj)
-        )
-
-    tkl_readings = []
-    for tkl_id, tkl_sensor in tkl_cache.items():
-        tkl_obj = TekelekSensor(**tkl_sensor)
-        tkl_readings.append(
-            tkl_to_bs(tkl_obj)
-            )
+    bb_readings = get_bb_readings(bb_cache)
+    tkl_readings = get_tkl_readings(tkl_cache)
 
     latest_readings = bb_readings + tkl_readings
     # latest_readings = filter_nulls(latest_readings)
